separate Sup3rGen and Sup3rGan
