+++ sup3r/models/multi_step.py
@@ -8,7 +8,7 @@ from phygnn.layers.custom_layers import Sup3rAdder, Sup3rConcat
 
 import sup3r.models
 from sup3r.models.abstract import AbstractSup3rGan
-from sup3r.models.base import Sup3rGan
+from sup3r.models.generator_discriminator import Sup3rGan
 
 
 logger = logging.getLogger(__name__)
